ENH: Add Test codes for itkBenchmarks
 - itkHighPriorityRealTimeProbeTest: to test the new methods of
   itkHighPriorityRealTimeProbe

- itkTimeProbeTest2: for the testing of the new itkResourceProbe

- itkHighPriorityRealTimeProbesCollectorTest: to test the functions of
  itkHighPriorityRealTimeProbesCollector.

- itkTimeProbesTest2: testing new Report functions of
  itkResourceProbesCollectorBase
